update room
- you can now create a room without a name
- some code refactoring
- delete some room typs
